@@ -1,23 +1,23 @@
 {
     "name": "justbetter/magento2-akeneo-bundle",
     "description": "Magento2 bundle for extending the Akeneo connector with awesome features.",
+    "license": "GPL-3.0-or-later",
     "require": {
-        "php": ">=8.0",
-        "akeneo/module-magento2-connector-community": "*"
+        "php": "^8.2",
+        "akeneo/module-magento2-connector-community": "^105.1.0"
     },
     "require-dev": {
         "bitexpert/phpstan-magento": "^0.11.0",
-        "phpstan/phpstan": "^1.10"
+        "phpstan/phpstan": "^1.10",
+        "phpro/grumphp": "^2.0",
+        "magento/magento-coding-standard": "^33",
+        "squizlabs/php_codesniffer": "^3.0"
     },
     "type": "magento2-module",
     "authors": [
         {
-            "name": "Pim Ruiter",
-            "email": "pim@justbetter.nl"
-        },
-        {
-            "name": "Robin Mulder",
-            "email": "robin@justbetter.nl"
+            "name": "JustBetter B.V.",
+            "email": "hello@justbetter.nl"
         }
     ],
     "repositories": {
@@ -37,7 +37,21 @@
     "config": {
         "allow-plugins": {
             "php-http/discovery": true,
-            "magento/composer-dependency-version-audit-plugin": true
+            "magento/composer-dependency-version-audit-plugin": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "phpro/grumphp": true
         }
+    },
+    "scripts": {
+        "analyse": "phpstan analyse --memory-limit=1G",
+        "codestyle": "vendor/bin/phpcs",
+        "codestyle:fix": "vendor/bin/phpcbf",
+        "grumphp": "grumphp run"
+    },
+    "scripts-descriptions": {
+        "analyse": "Run PHPStan static analysis",
+        "codestyle": "Check code style with PHP_CodeSniffer",
+        "codestyle:fix": "Automatically fix code style issues",
+        "grumphp": "Run all GrumPHP tasks (phpcs + phpstan)"
     }
 }
