@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace JustBetter\AkeneoBundle\Data;
 
@@ -7,21 +8,19 @@ use Magento\Framework\Data\OptionSourceInterface;
 
 class FamilyOptions implements OptionSourceInterface
 {
-    protected CollectionFactory $collectionFactory;
-
     public function __construct(
-        CollectionFactory $collectionFactory
+        protected CollectionFactory $collectionFactory // @phpstan-ignore-line
     ) {
-        $this->collectionFactory = $collectionFactory;
     }
 
+    /**
+     * @return array<int, array{value: int|string|null, label: string|null}>
+     */
     public function toOptionArray(): array
     {
-        return array_map(function ($set) {
-            return [
-                'value' => $set->getData('attribute_set_id'),
-                'label' => $set->getData('attribute_set_name')
-            ];
-        }, $this->collectionFactory->create()->getItems());
+        return array_map(fn ($set) => [
+            'value' => $set->getData('attribute_set_id'),
+            'label' => $set->getData('attribute_set_name'),
+        ], $this->collectionFactory->create()->getItems()); // @phpstan-ignore-line
     }
 }
