@@ -1,21 +1,30 @@
 <?php
+declare(strict_types=1);
 
 namespace JustBetter\AkeneoBundle\Plugin\Helper\Import;
 
+use Akeneo\Connector\Helper\Import\Product as ProductHelper;
 use Akeneo\Connector\Helper\Store as StoreHelper;
 use Magento\Framework\App\Config\ScopeConfigInterface;
 
 class Product
 {
-    protected $codes = null;
+    /**
+     * @var array<int, string>|null
+     */
+    protected ?array $codes = null;
 
     public function __construct(
         protected ScopeConfigInterface $config,
-        protected StoreHelper $storeHelper,
+        protected StoreHelper $storeHelper
     ) {
     }
 
-    public function beforeCreateTmpTableFromApi($subject, $result, $tableSuffix, mixed $family = null)
+    /**
+     * @param array<string, mixed> $result
+     * @return array{0: array<string, mixed>, 1: string, 2: mixed}
+     */
+    public function beforeCreateTmpTableFromApi(ProductHelper $subject, array $result, string $tableSuffix, mixed $family = null): array
     {
         if (is_null($this->codes)) {
             $this->codes = explode(',', (string)$this->config->getValue('akeneo_connector/justbetter/important_attributes'));
@@ -31,7 +40,7 @@ class Product
             foreach ($affectedStores as $affectedStore) {
                 $storeCodes[$affectedStore['lang'] . '-' . $affectedStore['channel_code']] = [
                     $affectedStore['lang'],
-                    $affectedStore['channel_code']
+                    $affectedStore['channel_code'],
                 ];
             }
         }
@@ -41,9 +50,9 @@ class Product
                 continue;
             }
             $result['values'][$code] = [[
-                    'locale' => null,
-                    'scope' => null,
-                    'data' => null,
+                'locale' => null,
+                'scope' => null,
+                'data' => null,
             ]];
 
             foreach ($storeCodes as $store) {
