@@ -1,20 +1,16 @@
 <?php
+declare(strict_types=1);
 
 namespace JustBetter\AkeneoBundle\Helper;
 
-use Magento\Store\Model\ScopeInterface;
 use Magento\Framework\App\Helper\AbstractHelper;
+use Magento\Store\Model\ScopeInterface;
 
 class SlackHelper extends AbstractHelper
 {
-    const XML_PATH = 'akeneo_connector/justbetter/slack/';
+    public const XML_PATH = 'akeneo_connector/justbetter/slack/';
 
-    /**
-     * getConfigValue function
-     * @param  String $field
-     * @param  Int $storeId
-     */
-    public function getConfigValue($field, int|string|null $storeId = null)
+    public function getConfigValue(string $field, ?int $storeId = null): mixed
     {
         return $this->scopeConfig->getValue(
             $field,
@@ -23,22 +19,13 @@ class SlackHelper extends AbstractHelper
         );
     }
 
-    /**
-     * getGeneralConfig function
-     * @param  String $code
-     * @param  null $storeId
-     */
-    public function getGeneralConfig($code, int|string|null $storeId = null)
+    public function getGeneralConfig(string $code, ?int $storeId = null): mixed
     {
         return $this->getConfigValue(self::XML_PATH . $code, $storeId);
     }
 
-    /**
-     * isEnable function
-     * @return boolean
-     */
-    public function isEnable()
+    public function isEnable(): bool
     {
-        return $this->getGeneralConfig('enable');
+        return (bool)$this->getGeneralConfig('enable');
     }
 }
