@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace JustBetter\AkeneoBundle\Model\ResourceModel;
 
@@ -6,12 +7,7 @@ use Magento\Framework\Model\ResourceModel\Db\AbstractDb;
 
 class Akeneo extends AbstractDb
 {
-    /**
-     * Initialize resource model
-     *
-     * @return void
-     */
-    protected function _construct()
+    protected function _construct(): void
     {
         $this->_init('akeneo_connector_entities', 'id');
     }
