@@ -1,18 +1,17 @@
 <?php
+declare(strict_types=1);
 
 namespace JustBetter\AkeneoBundle\Observer;
 
+use JustBetter\AkeneoBundle\Job\SetNotVisible as SetNotVisibleJob;
 use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
-use JustBetter\AkeneoBundle\Job\SetNotVisible as SetNotVisibleJob;
 
 class SetNotVisible implements ObserverInterface
 {
-    protected SetNotVisibleJob $job;
-
-    public function __construct(SetNotVisibleJob $job)
-    {
-        $this->job = $job;
+    public function __construct(
+        protected SetNotVisibleJob $job
+    ) {
     }
 
     public function execute(Observer $observer): void
