@@ -1,6 +1,7 @@
 <?php
+declare(strict_types=1);
 
-namespace JustBetter\AkeneoBundle\Plugin;
+namespace JustBetter\AkeneoBundle\Service;
 
 use Akeneo\Connector\Helper\Import\Entities;
 use Akeneo\Connector\Helper\Store as StoreHelper;
@@ -16,7 +17,7 @@ class CategoryExist
     ) {
     }
 
-    public function beforeSetValues()
+    public function execute(): void
     {
         $extensionEnabled = $this->config->getValue('akeneo_connector/justbetter/categoryexist', ScopeInterface::SCOPE_WEBSITE);
         if (!$extensionEnabled) {
@@ -24,12 +25,12 @@ class CategoryExist
         }
 
         $connection = $this->entitiesHelper->getConnection();
-
         $stores = $this->storeHelper->getStores('lang');
+
         foreach ($stores as $local => $affected) {
             foreach ($affected as $store) {
                 $columnName = 'url_key-' . $store['lang'];
-                
+
                 $query = "
                     UPDATE tmp_akeneo_connector_entities_category temp
                     LEFT JOIN catalog_category_entity_varchar eav ON (
@@ -43,14 +44,14 @@ class CategoryExist
                                 FROM eav_entity_type 
                                 WHERE entity_type_code = 'catalog_category'
                             )
-                            AND eav.`value` != ''
+                            AND eav.value != ''
                         )
                         AND eav.store_id = {$store['store_id']}
                     )
                     SET temp.`{$columnName}` = eav.value
                     WHERE temp._is_new = 0 AND eav.value IS NOT NULL
                 ";
-                
+
                 $connection->query($query);
             }
         }
