@@ -1,6 +1,6 @@
 <?php
-
-use \Magento\Framework\Component\ComponentRegistrar;
+declare(strict_types=1);
+use Magento\Framework\Component\ComponentRegistrar;
 
 ComponentRegistrar::register(
     ComponentRegistrar::MODULE,
